refactor: rename utils
